feat: refactor, clean
