+++ src/components/cards/WelcomeCard.vue
@@ -29,7 +29,7 @@
 <script lang="ts" setup>
 import { ref } from "vue";
 
-import BaseButton from "@/components/base/buttons/BaseButton.vue";
+import BaseButton from "@/components/base/BaseButton.vue";
 import BaseCard from "@/components/BaseCard.vue";
 import MessageDialog from "@/components/dialogs/MessageDialog.vue";
 import { lang } from "@/settings/lang";
