--- src/data/documents.js
@@ -1,114 +1,0 @@
-export default [
-  {
-    items: [
-      {
-        title: "Диплом",
-      },
-      {
-        label: "Квалификация",
-        text: "Психолог.<br/> Преподаватель психологии.",
-      },
-      {
-        label: "Специальность",
-        text: "Психология",
-      },
-      {
-        label: "Период обучения",
-        text: "2000 - 2005",
-      },
-    ],
-    image: "1.jpeg",
-  },
-  {
-    items: [
-      {
-        title: "Свидетельство о повышении квалификации",
-      },
-      {
-        label: "Программа",
-        text: "Гешталь-подход в области психологического консультирования",
-      },
-      {
-        label: "Год",
-        text: "2011",
-      },
-    ],
-    image: "2.jpeg",
-  },
-  {
-    items: [
-      {
-        title: "Диплом о профессио&shy;наль&shy;ной переподготовке",
-      },
-      {
-        label: "Программа",
-        text: "Современные методы психологической коррекции и консультирование",
-      },
-      {
-        label: "Период обучения",
-        text: "2005 - 2008",
-      },
-    ],
-    image: "3.jpeg",
-  },
-  {
-    items: [
-      {
-        title: "Свидетельство о повышении квалификации",
-      },
-      {
-        label: "Программа",
-        text: "Арт-терапия: разнообразие подходов",
-      },
-      {
-        label: "Количество часов",
-        text: "136",
-      },
-      {
-        label: "Период обучения",
-        text: "2010 - 2011",
-      },
-    ],
-    image: "4.jpeg",
-  },
-  {
-    items: [
-      {
-        title: "Удостоверение о повышении квалификации",
-      },
-      {
-        label: "Программа",
-        text: "Психологическое консультирование и психотерапия клиентов с сексуальными дисфункциями",
-      },
-      {
-        label: "Количество часов",
-        text: "72",
-      },
-      {
-        label: "Год",
-        text: "2013",
-      },
-    ],
-    image: "5.jpeg",
-  },
-  {
-    items: [
-      {
-        title: 'Удостоверение о краткосрочном повышении <span class="text-no-wrap">квалификации</span>',
-      },
-      {
-        label: "Программа",
-        text: "Снятие учебного и рабочего стресса",
-      },
-      {
-        label: "Количество часов",
-        text: "144",
-      },
-      {
-        label: "Год",
-        text: "2010",
-      },
-    ],
-    image: "6.jpeg",
-  },
-];
