--- src/helpers/formValidators.js
@@ -1,12 +1,0 @@
-export const formValidators = {
-  requiredName(value) {
-    return (value && !!value.trim()) || "Укажите Ваше имя";
-  },
-  requiredPhone(value) {
-    return (value && !!value.trim()) || "Укажите Ваш телефон";
-  },
-
-  requiredConnector(value) {
-    return (value && !!value.trim()) || "Выберете способ связи";
-  },
-};
