--- src/helpers/utils.ts
@@ -1,7 +1,0 @@
-export const isTestEnvironment = () => {
-  if (!localStorage) {
-    return false;
-  } else {
-    return localStorage.getItem("test_env") !== null;
-  }
-};
