--- src/settings/api.ts
@@ -1,3 +1,0 @@
-export const apiRoutes = {
-  telegramProcessor: "/telegramProcessor",
-};
