+++ src/settings/apiRoutes.ts
@@ -1,0 +1,3 @@
+export const apiRoutes = {
+  telegramProcessor: "/telegramProcessor",
+};
