--- src/settings/dates.ts
@@ -1,1 +1,0 @@
-export const DATE_TIME_FORMAT = "dd/MM/yyyy HH:mm O";
