--- src/settings/index.ts
@@ -1,12 +1,0 @@
-export const socialNetworks = [
-  {
-    link: "https://www.facebook.com/psysreda",
-    icon: "mdi-facebook",
-    color: "#4267B2",
-  },
-  {
-    link: "https://www.instagram.com/mypsyholog",
-    icon: "mdi-instagram",
-    color: "#e95950",
-  },
-];
