+++ time-tracking.txt
@@ -1,4 +1,4 @@
-20-Apr-2024 30m
+20-Apr-2024 40m
 18-Apr-2024 30m
 17-Apr-2024 2h
 15-Apr-2024 1h
